Add `NodeInputField` component
